rename to decls
