+++ lib/decls.js
