+++ lib/renderTag.js
@@ -12,7 +12,7 @@ export default function(classNameBuilder) {
                     cls)}
                 {...attrs}>
                     {content}
-            </Tag>;
+            </Tag>
         );
     };
 }
